Hooked up append/edit/delete actions to the item list.
git-svn-id: https://open-delta.googlecode.com/svn/trunk@509 dd429c29-c832-59a0-e864-46e68368aa0b

+++ delta-editor/src/main/java/au/org/ala/delta/editor/ItemController.java
@@ -1,12 +1,20 @@
 package au.org.ala.delta.editor;
 
+import java.awt.event.ActionEvent;
+
+import javax.swing.ActionMap;
 import javax.swing.JOptionPane;
+import javax.swing.JPopupMenu;
 
 import org.apache.commons.lang.StringUtils;
 import org.jdesktop.application.Action;
+import org.jdesktop.application.Application;
+import org.jdesktop.application.ApplicationContext;
 
 import au.org.ala.delta.editor.ui.EditorDataModel;
 import au.org.ala.delta.editor.ui.ItemList;
+import au.org.ala.delta.editor.ui.util.MenuBuilder;
+import au.org.ala.delta.editor.ui.util.PopupMenuListener;
 import au.org.ala.delta.model.Item;
 import au.org.ala.delta.ui.MessageDialogHelper;
 
@@ -17,50 +25,85 @@ public class ItemController {
 
 	private ItemList _view;
 	private EditorDataModel _model;
+	private ApplicationContext _context;
 	
 	public ItemController(ItemList view, EditorDataModel model) {
 		_view = view;
 		_model = model;
+		_context = Application.getInstance().getContext();
+		
+		buildPopup();
 	}
 	
+	public void buildPopup() {
+		JPopupMenu popup = new JPopupMenu();
+		ActionMap itemActions = _context.getActionMap(ItemController.class, this);
+		String[] itemsPopupActions = new String[] {"editItem", "addItem", "insertItem", "deleteItem", "-"};
+		MenuBuilder.buildMenu(popup, itemsPopupActions, itemActions);
+		
+		new PopupMenuListener(popup, _view);
+	}
+	
+	/**
+	 * Adds a new Item to the dataset after the last existing one.
+	 */
 	@Action
-	public void addItem() {
+	public void addItem(ActionEvent e) {
 		Item newItem = _model.addItem();
-		displayItemEditor(newItem.getItemNumber());
 		
-		if (StringUtils.isEmpty(newItem.getDescription())) {
-			//_model.deleteItem(newItem);
-		}
-		
+		editNewItem(newItem, e);
 	}
 	
+	/**
+	 * Deletes the currently selected Item.
+	 */
 	@Action
 	public void deleteItem() {
-		Item toDelete = getSelectedItem();
+		int selectedIndex = _view.getSelectedIndex();
+		Item toDelete = _view.getSelectedItem();
+		if (toDelete == null) {
+			return;
+		}
 		if (confirmDelete(toDelete)) {
 			_model.deleteItem(toDelete);
 		}
+		if (selectedIndex >= _model.getMaximumNumberOfItems()) {
+			selectedIndex = _model.getMaximumNumberOfItems()-1;
+		}
+		_view.setSelectedIndex(selectedIndex);
 	}
 	
+	/**
+	 * Displays the Item editor for the currently selected Item.
+	 */
 	@Action
-	public void editItem() {
+	public void editItem(ActionEvent e) {
+		_context.getActionMap().get("viewTaxonEditor").actionPerformed(e);
 		
 	}
 	
 	@Action
-	public void insertItem() {
+	public void insertItem(ActionEvent e) {
 		int selectedItem = _view.getSelectedIndex()+1;
-		//Item newItem = _model.insertItem(selectedItem+1);
-	}
-	
-	
-	private void displayItemEditor(int itemNumber) {
+		Item newItem = _model.addItem(selectedItem+1);
 		
+		editNewItem(newItem, e);
 	}
 	
-	public Item getSelectedItem() {
-		int itemNumber = _view.getSelectedIndex()+1;
-		return _model.getItem(itemNumber);
+	
+	public void editNewItem(Item newItem, ActionEvent e) {
+		int selectedIndex = newItem.getItemNumber()-1;
+		_view.setSelectedIndex(selectedIndex);
+		_view.ensureIndexIsVisible(selectedIndex);
+		editItem(e);
+		
+		if (StringUtils.isEmpty(newItem.getDescription())) {
+			_model.deleteItem(newItem);
+			if (selectedIndex >= _model.getMaximumNumberOfItems()) {
+				selectedIndex = _model.getMaximumNumberOfItems()-1;
+			}
+			_view.setSelectedIndex(selectedIndex);
+		}
 	}
 	
 	private boolean confirmDelete(Item toDelete) {
